CoreMigration - move sensor services to async_setup rather than async_setup_entry
